+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `file.contents` no longer renders YAML frontmatter
+
 ## [0.3.3] - 2026-09-06
 
 ### Changed